+++ CHANGELOG.md
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 
+## [Unreleased]
+
+### Added
+
+- Add table for page codes `Katakana`, `PC850` and `PC863`
+
 ## `0.10.0` (2024-04-09) [CURRENT]
 
 ### Added